Refactored all usages of int to int32_t/uint32_t or the 16-bit equivalent depending on usage context.

+++ onnx_import/code_templates/pico_cnn_average_pool1d.c
@@ -1,6 +1,6 @@
 {% if padding_needed %}
-const int {{identifier}}_padding[2] = { {{padding.0}}, {{padding.1}} };
-for (int i = 0; i < {{num_input_channels}}; i++) {
+const uint16_t {{identifier}}_padding[2] = { {{padding.0}}, {{padding.1}} };
+for (uint32_t i = 0; i < {{num_input_channels}}; i++) {
     average_pooling1d_naive_padded(({{input_buffer.name}}[i]),
                                    {{input_width}}, {{output_buffer.name}}[i],
                                    {{kernel_size}}, {{kernel_stride}},
@@ -13,7 +13,7 @@ for (int i = 0; i < {{num_input_channels}}; i++) {
                                    {{count_include_pad}});
 }
 {% else %}
-for (int i = 0; i < {{num_input_channels}}; i++) {
+for (uint32_t i = 0; i < {{num_input_channels}}; i++) {
     average_pooling1d_naive(({{input_buffer.name}}[i]),
                             {{input_width}}, {{output_buffer.name}}[i],
                             {{kernel_size}}, {{kernel_stride}},
