+++ pico-cnn/layers/pooling.c
@@ -28,7 +28,7 @@ void max_pooling1d_naive(const fp_t* input_channel, const uint16_t input_width,
 
 void max_pooling1d_naive_padded(const fp_t* input_channel, const uint16_t input_width,
                                 fp_t* output_channel, const uint16_t kernel_size, const uint16_t stride,
-                                const int* padding) {
+                                const uint16_t* padding) {
     fp_t* new_input_channel;
 
     extend_1d_input_with_padding(input_channel, input_width, &new_input_channel, padding, FLOAT_MIN);
@@ -79,7 +79,7 @@ void max_pooling2d_naive(const fp_t* input_channel, const uint16_t height, const
 
 void max_pooling2d_naive_padded(const fp_t* input_channel, const uint16_t height, const uint16_t width,
                                 fp_t* output_channel, const uint16_t kernel_size, const uint16_t stride,
-                                const int* padding) {
+                                const uint16_t* padding) {
 
     fp_t* new_input_channel;
     extend_2d_input_with_padding(input_channel, height, width, &new_input_channel, padding, FLOAT_MIN);
@@ -162,8 +162,8 @@ void average_pooling2d_naive(const fp_t* input_channel, const uint16_t height, c
                 // edge case
                 } else {
 
-                    int up_row, down_row, left_col, right_col;
-                    int divisor, div_row, div_col;
+                    uint16_t up_row, down_row, left_col, right_col;
+                    int32_t divisor, div_row, div_col;
 
                     up_row = MAX(channel_row, crop);
                     down_row = MIN(channel_row+kernel_size-1, height-crop-1);
@@ -197,7 +197,7 @@ void average_pooling2d_naive(const fp_t* input_channel, const uint16_t height, c
 
 void average_pooling2d_naive_padded(const fp_t* input_channel, const uint16_t height, const uint16_t width,
                                     fp_t* output_channel, const uint16_t kernel_size, const uint16_t stride,
-                                    fp_t bias, const int* padding, const uint16_t count_include_pad) {
+                                    fp_t bias, const uint16_t* padding, const uint16_t count_include_pad) {
 
     fp_t* new_input_channel;
     extend_2d_input_with_padding(input_channel, height, width, &new_input_channel, padding, 0.0);
@@ -258,8 +258,8 @@ void average_pooling1d_naive(const fp_t* input_channel, const uint16_t input_wid
              // edge case
             } else {
 
-                int left_col, right_col;
-                int divisor, div_col;
+                uint16_t left_col, right_col;
+                int32_t divisor, div_col;
 
                 left_col = MAX(input_channel_idx, crop);
                 right_col = MIN(input_channel_idx+kernel_size-1, input_width-crop-1);
@@ -278,8 +278,8 @@ void average_pooling1d_naive(const fp_t* input_channel, const uint16_t input_wid
 }
 
 void average_pooling1d_naive_padded(const fp_t* input_channel, const uint16_t input_width, fp_t* output_channel,
-                                    const uint16_t kernel_size, const uint16_t stride, fp_t bias, const int* padding,
-                                    const uint16_t count_include_pad) {
+                                    const uint16_t kernel_size, const uint16_t stride, fp_t bias,
+                                    const uint16_t* padding, const uint16_t count_include_pad) {
     fp_t* new_input_channel;
 
     extend_1d_input_with_padding(input_channel, input_width, &new_input_channel, padding, 0.0);
@@ -320,14 +320,14 @@ void global_max_pooling2d_naive(const fp_t* input_channel, const uint16_t input_
 }
 
 void pad_2d_naive(const fp_t* input_channel, const uint16_t height, const uint16_t width,
-                  fp_t* output_channel, const int* padding, fp_t initializer) {
+                  fp_t* output_channel, const uint16_t* padding, fp_t initializer) {
 
     uint16_t height_padded = height + padding[0] + padding[2];
     uint16_t width_padded = width + padding[1] + padding[3];
 
 
-    for(int r = 0; r < height_padded; r++){
-        for(int c = 0; c < width_padded; c++){
+    for(uint16_t r = 0; r < height_padded; r++){
+        for(uint16_t c = 0; c < width_padded; c++){
             output_channel[r*width_padded+c] = initializer;
         }
     }
@@ -339,10 +339,10 @@ void pad_2d_naive(const fp_t* input_channel, const uint16_t height, const uint16
 }
 
 void pad_1d_naive(const fp_t* input_channel, const uint16_t width,
-                  fp_t* output_channel, const int* padding, fp_t initializer) {
+                  fp_t* output_channel, const uint16_t* padding, fp_t initializer) {
     uint16_t width_padded = width + padding[0] + padding[1];
 
-    for(int i = 0; i < width_padded; i++){
+    for(uint16_t i = 0; i < width_padded; i++){
         output_channel[i] = initializer;
     }
